requirements.txt
Django==2.1.2
channels==2.1.3
python_dateutil==2.7.3
djangorestframework==3.8.2
